And yet another fun Mandelbrot project. Completely refactored the code to speed things up. Now it's possible to have a live preview in the zoombox.
The new method separates the calculations from the drawing, making it possible to swap the color palette without doing all the recalculations. So, now the slider for the palette immediately sends a request to calculate a new palette and immediately apply it to the current situation.

When zooming in the previous situation is stored in a history list. So, using the right mouse-button you can quickly go back to your previous situation, even preserving the palette.